processor_py/requirements.txt
simplejson
gevent
pyzmq
hotqueue
